Backjoon, Book1 예제 추가

+++ Backjoon/src/step01/_06.java
@@ -1,12 +1,16 @@
 package step01;
-
+/*
+ * 날짜: 2022/08/19
+ * 이름: 김재준
+ * 내용: 백준 1단계 6번 문제. A/B
+ */
 import java.util.Scanner;
 
 public class _06 {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
-		int a = sc.nextInt();
-		int b = sc.nextInt();
+		double a = sc.nextInt();
+		double b = sc.nextInt();
 		System.out.println(a/b);
 		sc.close();
 	}
